+++ src/Mod/Draft/draftgeoutils/__init__.py
@@ -1,0 +1,40 @@
+# ***************************************************************************
+# *   Copyright (c) 2009, 2010 Yorik van Havre <yorik@uncreated.net>        *
+# *   Copyright (c) 2009, 2010 Ken Cline <cline@frii.com>                   *
+# *                                                                         *
+# *   This file is part of the FreeCAD CAx development system.              *
+# *                                                                         *
+# *   This program is free software; you can redistribute it and/or modify  *
+# *   it under the terms of the GNU Lesser General Public License (LGPL)    *
+# *   as published by the Free Software Foundation; either version 2 of     *
+# *   the License, or (at your option) any later version.                   *
+# *   for detail see the LICENCE text file.                                 *
+# *                                                                         *
+# *   FreeCAD is distributed in the hope that it will be useful,            *
+# *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+# *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+# *   GNU Library General Public License for more details.                  *
+# *                                                                         *
+# *   You should have received a copy of the GNU Library General Public     *
+# *   License along with FreeCAD; if not, write to the Free Software        *
+# *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+# *   USA                                                                   *
+# *                                                                         *
+# ***************************************************************************
+"""Modules that contain functions that use and manipulate shapes.
+
+These functions provide support for dealing with the custom objects
+defined within the workbench.
+The functions are meant to be used in the creation step of the objects,
+or to manipulate the created shapes, principally by the functions
+in the `draftmake` package.
+
+These functions should deal with the internal shapes of the objects,
+and their special properties; they shouldn't be very generic.
+
+These functions may be useful for other programmers in their own macros
+or workbenches. These functions may not necessarily be exposed as
+part of the Draft workbench programming interface yet.
+
+These functions were previously defined in the big `DraftGeomUtils` module.
+"""
